add function and refactor code

+++ src/main/java/com/khaikin/qrest/combo/ComboRepository.java
@@ -2,5 +2,5 @@ package com.khaikin.qrest.combo;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ComboRepository extends JpaRepository<Combo, Integer> {
+public interface ComboRepository extends JpaRepository<Combo, Long> {
 }
