+++ src/main/java/com/khaikin/qrest/comboitem/ComboItemRepository.java
@@ -2,5 +2,5 @@ package com.khaikin.qrest.comboitem;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ComboItemRepository extends JpaRepository<ComboItem, Integer> {
+public interface ComboItemRepository extends JpaRepository<ComboItem, Long> {
 }
